+++ src/main/java/study/clone/api/RestApiController.java
@@ -1,0 +1,13 @@
+package study.clone.api;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class RestApiController {
+
+  @GetMapping("home")
+  public String home() {
+    return "<h1>home<h1>";
+  }
+}
