+++ src/main/java/study/clone/config/old/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package study.clone.config.jwt;
+package study.clone.config.old.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
