+++ src/main/java/study/clone/config/old/jwt/JwtToken.java
@@ -1,0 +1,17 @@
+package study.clone.config.old.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Builder
+@Setter
+@Getter
+@AllArgsConstructor
+public class JwtToken {
+
+  private String grantType;
+  private String accessToken;
+  private String refreshToken;
+}
