+++ src/main/java/study/clone/service/MemberService.java
@@ -8,8 +8,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import study.clone.config.jwt.JwtToken;
-import study.clone.config.jwt.JwtTokenProvider;
+import study.clone.config.old.jwt.JwtToken;
+import study.clone.config.old.jwt.JwtTokenProvider;
 import study.clone.dto.MemberDto;
 import study.clone.entity.Member;
 import study.clone.exception.DuplicateUsernameException;
